+++ cyvcf2/helpers.h
@@ -7,4 +7,5 @@ int gt_types_012_from_int8(const int8_t *data, int num_samples, int ploidy,
                            int strict_gt, int HOM_ALT, int UNKNOWN,
                            int32_t *gt_types, int32_t *gt_idxs,
                            int32_t *gt_phased);
+int vcf_line_strip_format(char *line, int len, const char *keep);
 int32_t* bcf_hdr_seqlen(const bcf_hdr_t *hdr, int32_t *nseq);